refactoring to library
