+++ Cargo.toml
@@ -27,3 +27,5 @@ strum_macros = "0.20.1"
 strum = "0.20.0"
 dirs = "3.0.1"
 serde_derive = "1.0.118"
+#github = "0.1.2"
+semver = "0.11.0"
